fix: temporarily remove billing address fields
Payrix expects 3-letter ISO code for the country, and for state it expects either a 2 character postal abbreviation when in the U.S. or if outside of the U.S. the full state name.

The current implementation cannot guarantee nor test for this, so just don't send it all until we can!